Suppress errors if page number frame not found
The left and right frames are fetched separately, so if one page is numbered it still appears
in the page prompt even if the other doesn't have one.

It checks that the fetched page number can be converted to a number, with the intention of
catching placeholders (such as "X").