Trinity PDF: Always create separate pages from spreads
This was added to the script ages ago, but for some reason
I didn't commit it to the repository. Had to salvage it
from someone else's machine as the date-checking version
was based on the old revision of the script which exported
real spreads.

Exporting real spreads causing a problem when sending the
paper to outside providers who need single-page PDFs.