Trinity: Update for CC & fix bugs in date-checking
In particular, L- and R-Edition date were phased out a while
ago (all date frames are now just "Edition date").

I’ve also introduced master-checking for dates as the
frame doesn’t appear on the "real" pages to the script
if it’s inherited from the master (I think this is a post-CS4
change). Now, we check the "real" page, then its master,
then the master's master, until we reach the root master.